testing/web-platform/tests/web-animations/interfaces/AnimationEffectTiming/direction.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>AnimationEffectTiming.direction</title>
<link rel="help" href="https://w3c.github.io/web-animations/#dom-animationeffecttiming-direction">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../testcommon.js"></script>
<body>
<div id="log"></div>
<script>
'use strict';

test(function(t) {
  var anim = createDiv(t).animate(null);
  assert_equals(anim.effect.timing.direction, 'normal');
}, 'Has the default value \'normal\'');

test(function(t) {
  var div = createDiv(t);
  var anim = div.animate({ opacity: [ 0, 1 ] }, 2000);

  var directions = ['normal', 'reverse', 'alternate', 'alternate-reverse'];
  directions.forEach(function(direction) {
    anim.effect.timing.direction = direction;
    assert_equals(anim.effect.timing.direction, direction,
                  'set direction to ' + direction);
  });
}, 'Can be set to each of the possible keywords');

</script>
</body>
